requirement.txt
python-vlc
pyfiglet
clint
pyyaml